@@ -1,23 +1,21 @@
 {
   "private": true,
-  "version": "3.0.0-alpha.14",
+  "version": "3.0.0-alpha.14.4.0",
   "dependencies": {},
   "devDependencies": {
     "assert": "~1.3.0",
     "axios": "^0.18.0",
     "babel-eslint": "^6.1.2",
     "chalk": "^2.4.1",
-    "eslint": "^3.12.2",
+    "eslint": "^4.19.1",
     "eslint-plugin-babel": "^4.0.0",
-    "eslint-plugin-react": "^6.8.0",
-    "eslint-plugin-redux-saga": "^0.3.0",
-    "gitbook-cli": "^2.3.2",
+    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-redux-saga": "^0.8.0",
     "istanbul": "~0.4.2",
     "jest": "^22.4.3",
     "jest-cli": "^22.4.4",
     "lerna": "^2.0.0",
     "lodash": "^4.17.5",
-    "mocha": "~2.4.5",
     "pre-commit": "~1.1.2",
     "redux-saga": "^0.14.3",
     "request": "^2.87.0",
@@ -25,8 +23,8 @@
     "strapi-lint": "file:packages/strapi-lint"
   },
   "scripts": {
-    "clean": "npm run removesymlinkdependencies && rm -rf package-lock.json && rm -rf packages/*/package-lock.json",
-    "clean:all": "npm run removesymlinkdependencies && rm -rf package-lock.json && rm -rf packages/*/package-lock.json && rm -rf packages/*/node_modules",
+    "clean": "npm run removesymlinkdependencies && npx rimraf package-lock.json && npx rimraf packages/*/package-lock.json",
+    "clean:all": "npm run removesymlinkdependencies && npx rimraf package-lock.json && npx rimraf packages/*/package-lock.json && npx rimraf packages/*/node_modules",
     "doc": "node ./scripts/documentation.js",
     "release": "npm run clean:all && npm install && npm run createsymlinkdependencies && lerna exec --concurrency 1 -- npm install && npm run removesymlinkdependencies && node ./scripts/publish.js $TAG",
     "createsymlinkdependencies": "node ./scripts/createSymlinkDependencies.js",
