@@ -48,13 +48,13 @@ module.exports = scope => {
     },
     'devDependencies': {
       'babel-eslint': '^7.1.1',
-      'eslint': '^3.12.2',
+      'eslint': '^4.19.1',
       'eslint-config-airbnb': '^13.0.0',
-      'eslint-plugin-import': '^2.2.0',
-      'eslint-plugin-react': '^6.8.0'
+      'eslint-plugin-import': '^2.11.0',
+      'eslint-plugin-react': '^7.7.0'
     },
     'dependencies': Object.assign({}, {
-      'lodash': '4.x.x',
+      'lodash': '^4.17.5',
       'strapi': getDependencyVersion(cliPkg, 'strapi'),
       [scope.client.connector]: getDependencyVersion(cliPkg, 'strapi'),
     }, additionalsDependencies, {
