@@ -66,7 +66,7 @@ module.exports = (scope, cb) => {
   const connectionValidation = () => {
     const databaseChoices = [
       {
-        name: 'MongoDB (recommended)',
+        name: 'MongoDB',
         value: {
           database: 'mongo',
           connector: 'strapi-hook-mongoose'
@@ -152,7 +152,7 @@ module.exports = (scope, cb) => {
                   when: !hasDatabaseConfig,
                   type: 'input',
                   name: 'port',
-                  message: 'Port (It will be ignored if you enable +srv):',
+                  message: `Port${scope.client.database === 'mongo' ? ' (It will be ignored if you enable +srv)' : ''}:`,
                   default: (answers) => { // eslint-disable-line no-unused-vars
                     if (_.get(scope.database, 'port')) {
                       return scope.database.port;
@@ -186,7 +186,7 @@ module.exports = (scope, cb) => {
                   when: !hasDatabaseConfig && scope.client.database === 'mongo',
                   type: 'input',
                   name: 'authenticationDatabase',
-                  message: 'Authentication database:',
+                  message: 'Authentication database (Maybe "admin" or blank):',
                   default: _.get(scope.database, 'authenticationDatabase', undefined)
                 },
                 {
@@ -203,7 +203,7 @@ module.exports = (scope, cb) => {
                 }
 
                 scope.database.settings.host = answers.host;
-                scope.database.settings.srv = answers.srv;
+                scope.database.settings.srv = _.toString(answers.srv) === 'true';
                 scope.database.settings.port = answers.port;
                 scope.database.settings.database = answers.database;
                 scope.database.settings.username = answers.username;
@@ -225,14 +225,14 @@ module.exports = (scope, cb) => {
               shell.exec(`mkdir ${scope.tmpPath}`);
             }
 
-            let cmd = `${packageCmd} ${scope.client.connector}@alpha`;
+            let cmd = `${packageCmd} ${scope.client.connector}@${scope.strapiPackageJSON.version}`;
 
             if (scope.client.module) {
               cmd += ` ${scope.client.module}`;
             }
 
             if (scope.client.connector === 'strapi-hook-bookshelf') {
-              cmd += ` strapi-hook-knex@alpha`;
+              cmd += ` strapi-hook-knex@${scope.strapiPackageJSON.version}`;
 
               scope.additionalsDependencies = ['strapi-hook-knex', 'knex'];
             }
@@ -260,7 +260,7 @@ module.exports = (scope, cb) => {
             } catch(err) {
               shell.rm('-r', scope.tmpPath);
               console.log(err);
-              cb.success();
+              cb.error();
             }
           });
       });
