@@ -1,26 +1,16 @@
 /**
  * WEBPACK DLL GENERATOR
  *
- * This profile is used to cache webpack's module
- * contexts for external library and framework type
- * dependencies which will usually not change often enough
- * to warrant building them from scratch every time we use
- * the webpack process.
+ * This profile is used to cache webpack's module contexts for external library and framework type dependencies which
+ * will usually not change often enough to warrant building them from scratch every time we use the webpack process.
  */
 
 const path = require('path');
 const webpack = require('webpack');
 const isAdmin = process.env.IS_ADMIN === 'true';
 
-const appPath = (() => {
-  if (process.env.APP_PATH) {
-    return process.env.APP_PATH;
-  }
-
-  return isAdmin ? path.resolve(process.env.PWD, '..') : path.resolve(process.env.PWD, '..', '..');
-})();
-// const isSetup = path.resolve(process.env.PWD, '..', '..') === path.resolve(process.env.INIT_CWD);
-const isSetup = process.env.IS_MONOREPO;
+const isSetup = process.env.IS_MONOREPO || false;
+const appPath = process.env.APP_PATH || path.resolve(process.env.PWD, '..', ( isAdmin ? '' : '..' ));
 
 const rootAdminpath = (() => {
   if (isSetup) {
@@ -69,6 +59,7 @@ module.exports = {
       'reactstrap': path.resolve(rootAdminpath, 'node_modules', 'strapi-helper-plugin', 'node_modules', 'reactstrap'),
       'react-dnd': path.resolve(rootAdminpath, 'node_modules', 'strapi-helper-plugin', 'node_modules', 'react-dnd'),
       'react-dnd-hmtl5-backend': path.resolve(rootAdminpath, 'node_modules', 'strapi-helper-plugin', 'node_modules', 'react-dnd-html5-backend'),
+      'styled-components': path.resolve(rootAdminpath, 'node_modules', 'strapi-helper-plugin', 'node_modules', 'styled-components'),
     },
     symlinks: false,
     extensions: [
