@@ -1,11 +1,10 @@
 'use strict';
 
-// Node.js core.
-const execSync = require('child_process').execSync;
-const path = require('path');
+// Public node modules
+const rimraf = require('rimraf');
 
 module.exports = (scope, success, error) => {
-  const Mongoose = require(path.resolve(`${scope.tmpPath}/node_modules/mongoose`));
+  const Mongoose = require('mongoose');
 
   const { username, password, srv } = scope.database.settings;
   const { authenticationDatabase, ssl } = scope.database.options;
@@ -36,8 +35,11 @@ module.exports = (scope, success, error) => {
 
     Mongoose.connection.close();
 
-    execSync(`rm -r "${scope.tmpPath}"`);
-
-    success();
+    rimraf(scope.tmpPath, (err) => {
+      if (err) {
+        console.log(`Error removing connection test folder: ${scope.tmpPath}`);
+      }
+      success();
+    });
   });
 };
