@@ -1,14 +1,13 @@
 'use strict';
 
-/* eslint-disable import/no-unresolved */
-// Node.js core.
-const execSync = require('child_process').execSync;
+// Public node modules
+const rimraf = require('rimraf');
 
 // Logger.
 const logger = require('strapi-utils').logger;
 
 module.exports = (scope, success, error) => {
-  const Redis = require(`${scope.tmpPath}/node_modules/ioredis`);
+  const Redis = require(`ioredis`);
   const redis = new Redis({
     port: scope.database.settings.port,
     host: scope.database.settings.host,
@@ -26,10 +25,14 @@ module.exports = (scope, success, error) => {
 
     logger.info('The app has been connected to the database successfully!');
 
-    execSync(`rm -r "${scope.tmpPath}"`);
+    rimraf(scope.tmpPath, (err) => {
+      if (err) {
+        console.log(`Error removing connection test folder: ${scope.tmpPath}`);
+      }
+      logger.info('Copying the dashboard...');
 
-    logger.info('Copying the dashboard...');
+      success();
+    });
 
-    success();
   });
 };
