@@ -184,7 +184,7 @@ module.exports = {
         withRelated: populate || this.associations.map(x => x.alias)
       });
 
-    const data = record.toJSON ? record.toJSON() : record;
+    const data = _.get(record, 'toJSON()', record);
 
     // Retrieve data manually.
     if (_.isEmpty(populate)) {
