@@ -106,11 +106,12 @@ module.exports = {
 
     const request = await this.create(values)
       .catch((err) => {
-        const message = err.message.split('index:');
-        const field = _.words(_.last(message).split('_')[0]);
-        const error = { message: `This ${field} is already taken`, field };
-
-        throw error;
+        if (err.message) {
+          const message = err.message.split('index:');
+          const field = _.words(_.last(message).split('_')[0]);
+          err = { message: `This ${field} is already taken`, field };
+        }
+        throw err;
       });
 
     // Transform to JSON object.
