@@ -183,6 +183,18 @@ module.exports = {
       return ctx.badRequest(null, [{ messages: removeModelErrors }]);
     }
 
+    const pluginStore = strapi.store({
+      environment: '',
+      type: 'plugin',
+      name: 'content-manager'
+    });
+
+    const schema = await pluginStore.get({ key: 'schema' });
+
+    delete schema.layout[model];
+
+    await pluginStore.set({ key: 'schema', value: schema });
+
     ctx.send({ ok: true });
 
     strapi.reload();
@@ -190,7 +202,7 @@ module.exports = {
 
   autoReload: async ctx => {
     ctx.send({
-      autoReload: _.get(strapi.config.environments, 'development.server.autoReload', false),
+      autoReload: _.get(strapi.config.currentEnvironment, 'server.autoReload', { enabled: false })
     });
   },
 
