@@ -334,8 +334,8 @@ module.exports = {
 
   autoReload: async ctx => {
     ctx.send({
-      autoReload: _.get(strapi.config.environments, 'development.server.autoReload', false),
-      environment: strapi.config.environment,
+      autoReload: _.get(strapi.config.currentEnvironment, 'server.autoReload', { enabled: false }),
+      environment: strapi.config.environment
     });
   }
 };
