@@ -24,7 +24,7 @@ module.exports = {
     }
 
     // Extract optional relational data.
-    const { refId, ref, source, field } = ctx.request.body.fields;
+    const { refId, ref, source, field, path } = ctx.request.body.fields;
     const { files = {} } = ctx.request.body.files;
 
     if (_.isEmpty(files)) {
@@ -47,6 +47,13 @@ module.exports = {
             source,
             field
           }]
+        });
+      }
+
+      // Update uploading folder path for the file.
+      if (path) {
+        Object.assign(file, {
+          path
         });
       }
 
