@@ -70,10 +70,12 @@ module.exports = {
       upload: (file) => {
         return new Promise((resolve, reject) => {
           // upload file on S3 bucket
+          const path = file.path ? `${file.path}/` : '';
           S3.upload({
-            Key: `${file.hash}${file.ext}`,
+            Key: `${path}${file.hash}${file.ext}`,
             Body: new Buffer(file.buffer, 'binary'),
-            ACL: 'public-read'
+            ACL: 'public-read',
+            ContentType: file.mime,
           }, (err, data) => {
             if (err) {
               return reject(err);
@@ -89,12 +91,9 @@ module.exports = {
       delete: (file) => {
         return new Promise((resolve, reject) => {
           // delete file on S3 bucket
-          S3.deleteObjects({
-            Delete: {
-              Objects: [{
-                Key: `${file.hash}${file.ext}`
-              }]
-            }
+          const path = file.path ? `${file.path}/` : '';
+          S3.deleteObject({
+            Key: `${path}${file.hash}${file.ext}`
           }, (err, data) => {
             if (err) {
               return reject(err);
