@@ -12,7 +12,7 @@ const _ = require('lodash');
 
 // Following this discussion https://stackoverflow.com/questions/18082/validate-decimal-numbers-in-javascript-isnumeric this function is the best implem to determine if a value is a valid number candidate
 const isNumeric = (value) => {
-  return !isNaN(parseFloat(value)) && isFinite(value);
+  return !_.isObject(value) && !isNaN(parseFloat(value)) && isFinite(value);
 };
 
 /* eslint-disable prefer-template */
@@ -466,7 +466,7 @@ module.exports = {
         // Remove the filter keyword at the end
         let splitKey = key.split('_').slice(0,-1);
         splitKey = splitKey.join('_');
-        
+
         if (modelAttributes[splitKey]) {
           fieldType = modelAttributes[splitKey]['type'];
         }
