@@ -2,24 +2,23 @@ const path = require('path');
 const shell = require('shelljs');
 const chalk = require('chalk');
 const eslintErrorsFormatter = require('./eslintErrorsFormatter');
-const glob = require('glob');
-const fs = require('fs');
 const listChangedFiles = require('../packages/strapi-lint/lib/internals/shared/listChangedFiles.js');
 const changedFiles = listChangedFiles();
-const { includes, take } = require('lodash');
+const { take, template } = require('lodash');
 
-const frontCmd =
-  'node ../../node_modules/strapi-lint/node_modules/.bin/eslint --ignore-path .gitignore --ignore-pattern \'/admin/build/\' --config ../../node_modules/strapi-lint/lib/internals/eslint/front/.eslintrc.json admin';
-const helperCmd =
-  'node ../../node_modules/strapi-lint/node_modules/.bin/eslint --ignore-path .gitignore --ignore-pattern \'/admin/build/\' --config ../../node_modules/strapi-lint/lib/internals/eslint/front/.eslintrc.json lib/src';
-const backCmd =
-  'node ../../node_modules/strapi-lint/node_modules/.bin/eslint --ignore-path .gitignore --ignore-pattern \'/admin\' --config ../../node_modules/strapi-lint/lib/internals/eslint/back/.eslintrc.json controllers config services bin lib';
+const cmdEslint = template(
+  'node ../../node_modules/strapi-lint/node_modules/.bin/eslint --ignore-path .gitignore --ignore-pattern "${ignore}"'
+  + ' --config ../../node_modules/strapi-lint/lib/internals/eslint/${conf}/.eslintrc.json ${params}'
+);
 
+const cmdFront = cmdEslint({ ignore: '/admin/build/', conf: 'front', params: 'admin' });
+const cmdHelper = cmdEslint({ ignore: '/admin/build/', conf: 'front', params: 'lib/src' });
+const cmdBack = cmdEslint({ ignore: '/admin', conf: 'back', params: 'controllers config services bin lib' });
 
-const watcher = (label, pckgName, type = 'front') => {
+const watcher = (label, pckgName) => {
   shell.echo(label);
   shell.cd(pckgName);
-  const cmd = includes(pckgName, 'strapi-helper-plugin') ? helperCmd : `${frontCmd} && ${backCmd}`;
+  const cmd = pckgName.includes('strapi-helper-plugin') ? cmdHelper : `${cmdFront} && ${cmdBack}`;
 
   const data = shell.exec(cmd, { silent: true });
   shell.echo(chalk(eslintErrorsFormatter(data.stdout)));
@@ -31,19 +30,18 @@ const watcher = (label, pckgName, type = 'front') => {
   shell.echo('');
 };
 
-const files = glob
-  .sync('**/*.js', { ignore: '**/node_modules/**' })
-  .filter(f => changedFiles.has(f))
-  .filter(
-    package =>
-      !package.includes('README.md') &&
-      !package.includes('strapi-middleware-views') &&
-      !package.includes('strapi-lint') &&
-      !package.includes('strapi-plugin-settings-manager') &&
-      !package.includes('scripts') &&
-      !package.includes('test') &&
-      !package.includes('jest.config.js')
-  )
+const except = [
+  'docs',
+  'jest.config.js',
+  'scripts',
+  'strapi-lint',
+  'strapi-middleware-views',
+  'strapi-plugin-settings-manager',
+  'test',
+];
+
+const changedDirs = [...changedFiles]
+  .filter(file => path.extname(file) === '.js' && !except.some(path => file.includes(path)))
   .map(file => {
     const directoryArray = file.split('/');
     const toTake = directoryArray.length === 2 ? 1 : 2;
@@ -51,9 +49,7 @@ const files = glob
     return take(directoryArray, toTake).join('/');
   });
 
-
-files
-  .filter((directory, index) => files.indexOf(directory) === index)
-  .forEach(package => {
-    watcher(`Testing ${package}`, package);
+[...new Set(changedDirs)]
+  .forEach(directory => {
+    watcher(`Testing ${directory}`, directory);
   });
